Refactor DisclosureHeading into more semantic SectionHeading component
